feat: added delete product operations

+++ client/src/components/products/Edit.jsx
@@ -61,19 +61,21 @@ const Edit = () => {
   };
 
   const deleteCategory = async (id) => {
+   if(window.confirm('Ürünü silmek istediğinize emin misiniz?')){
     try {
-      fetch(`http://localhost:5000/api/categories/delete-category`, {
+      fetch(`http://localhost:5000/api/products/delete-product`, {
         method: 'DELETE',
-        body: JSON.stringify({ categoryId: id }),
+        body: JSON.stringify({ productId: id }),
         headers: { 'Content-type': 'application/json; charset=UTF-8' },
       });
-      message.success(` Kategori başarıyla silindi.`);
+      message.success(` Ürün başarıyla silindi.`);
      
-      setCategories(categories.filter((item) => item._id !== id));
+      setProducts(products.filter((item) => item._id !== id));
     } catch (error) {
       message.success(`Bir şeyler yanlış gitti.`);
       console.log(error);
     }
+   }
   };
   
   const CloseModalAll = () => {
